Refactor check-user-access function
Refactor the check-user-access Supabase function into smaller, more maintainable files.
