+++ supabase/functions/check-user-access/config.toml
@@ -1,0 +1,2 @@
+
+project_id = "jpelncawdaounkidvymu"
