+++ supabase/functions/check-user-access/index.ts
@@ -1,12 +1,12 @@
 
 import { serve } from "https://deno.land/std@0.190.0/http/server.ts"
 import { createClient } from "https://esm.sh/@supabase/supabase-js@2.45.0"
-
-const corsHeaders = {
-  'Access-Control-Allow-Origin': '*',
-  'Access-Control-Allow-Headers': 'authorization, x-client-info, apikey, content-type',
-  'Access-Control-Allow-Methods': 'GET, POST, OPTIONS',
-}
+import { corsHeaders } from "../_shared/cors.ts"
+import { authenticateUser, createResponse } from "./auth.ts"
+import { checkBetaAccess } from "./beta-subscription.ts"
+import { checkPaidAccess } from "./paid-subscription.ts"
+import { checkTrialAccess } from "./trial-subscription.ts"
+import { checkDevelopmentMode } from "./development.ts"
 
 serve(async (req) => {
   // Handle CORS preflight requests
@@ -21,313 +21,64 @@ serve(async (req) => {
   try {
     console.log("Starting check-user-access function");
     
-    // Récupérer le token d'authentification
+    // Get authentication token
     const authHeader = req.headers.get('Authorization');
     
-    if (!authHeader) {
-      console.log("No Authorization header found");
-      return new Response(
-        JSON.stringify({ 
-          access: false, 
-          message: 'Non authentifié' 
-        }),
-        { 
-          headers: { ...corsHeaders, 'Content-Type': 'application/json' },
-          status: 401 
-        }
-      );
-    }
-    
-    const token = authHeader.replace('Bearer ', '');
-    console.log("Token extracted from Authorization header");
-    
-    // Initialiser le client Supabase
+    // Initialize Supabase client
     const supabaseClient = createClient(
       Deno.env.get('SUPABASE_URL') ?? '',
       Deno.env.get('SUPABASE_ANON_KEY') ?? '',
     );
     console.log("Supabase client initialized");
 
-    // Vérifier l'utilisateur
-    console.log("Checking user with token");
-    const { data: { user }, error: userError } = await supabaseClient.auth.getUser(token);
-    
-    if (userError) {
-      console.error("User error:", userError);
-      return new Response(
-        JSON.stringify({ 
-          access: false, 
-          message: 'Utilisateur non trouvé',
-          error: userError.message
-        }),
-        { 
-          headers: { ...corsHeaders, 'Content-Type': 'application/json' },
-          status: 401 
-        }
-      );
+    // Authenticate user
+    const authResult = await authenticateUser(supabaseClient, authHeader);
+    if (authResult.error) {
+      return createResponse(authResult.body, authResult.status);
     }
     
-    if (!user) {
-      console.log("No user found");
-      return new Response(
-        JSON.stringify({ 
-          access: false, 
-          message: 'Utilisateur non trouvé' 
-        }),
-        { 
-          headers: { ...corsHeaders, 'Content-Type': 'application/json' },
-          status: 401 
-        }
-      );
+    const user = authResult.user;
+
+    // Check for development mode
+    const devModeResult = checkDevelopmentMode(Deno.env.get('ENVIRONMENT'));
+    if (devModeResult) {
+      return createResponse(devModeResult);
     }
+
+    // Check subscription access in priority order: Beta -> Paid -> Trial
     
-    console.log("User found:", user.email);
-
-    // En mode développement, retourner accès actif pour faciliter les tests
-    if (Deno.env.get('ENVIRONMENT') === 'development') {
-      console.log("Development environment detected, granting access");
-      return new Response(
-        JSON.stringify({ 
-          access: true, 
-          type: 'dev_mode',
-          expires_at: null,
-          message: 'Accès accordé en mode développement'
-        }),
-        { 
-          headers: { ...corsHeaders, 'Content-Type': 'application/json' },
-          status: 200 
-        }
-      );
+    // 1. Check for beta access (HIGHEST PRIORITY)
+    const betaResult = await checkBetaAccess(supabaseClient, user);
+    if (betaResult) {
+      return createResponse(betaResult);
     }
 
-    // Vérifier si c'est un utilisateur beta - PRIORITAIRE
-    console.log("Checking if user is beta user");
-    try {
-      // Première méthode: Vérifier dans la table beta_users
-      const { data: betaUser, error: betaError } = await supabaseClient
-        .from('beta_users')
-        .select('*')
-        .eq('email', user.email)
-        .single();
-        
-      if (betaError && betaError.code !== 'PGRST116') {
-        console.error("Beta user check error:", betaError);
-      }
-        
-      if (betaUser) {
-        console.log('Utilisateur beta trouvé dans la table beta_users');
-        return new Response(
-          JSON.stringify({ 
-            access: true, 
-            type: 'beta',
-            expires_at: null,
-          }),
-          { 
-            headers: { ...corsHeaders, 'Content-Type': 'application/json' },
-            status: 200 
-          }
-        );
-      }
-
-      // Deuxième méthode: Vérifier dans la table user_subscriptions avec type=beta
-      const { data: betaSubscription, error: betaSubError } = await supabaseClient
-        .from('user_subscriptions')
-        .select('*')
-        .eq('user_id', user.id)
-        .eq('type', 'beta')
-        .eq('status', 'active')
-        .order('created_at', { ascending: false })
-        .limit(1)
-        .maybeSingle();
-
-      if (betaSubError) {
-        console.error("Beta subscription check error:", betaSubError);
-      }
-
-      if (betaSubscription) {
-        console.log('Abonnement beta trouvé dans user_subscriptions:', betaSubscription);
-        
-        // Vérifier si l'abonnement beta est expiré
-        if (betaSubscription.expires_at) {
-          const expiryDate = new Date(betaSubscription.expires_at);
-          if (expiryDate < new Date()) {
-            console.log("Beta subscription expired at:", expiryDate);
-            return new Response(
-              JSON.stringify({ 
-                access: false, 
-                message: 'Accès beta expiré',
-                type: 'beta',
-                expires_at: betaSubscription.expires_at
-              }),
-              { 
-                headers: { ...corsHeaders, 'Content-Type': 'application/json' },
-                status: 200
-              }
-            );
-          }
-        }
-        
-        console.log("Active beta subscription found, granting access");
-        return new Response(
-          JSON.stringify({ 
-            access: true, 
-            type: 'beta',
-            expires_at: betaSubscription.expires_at,
-            promo_code: betaSubscription.promo_code
-          }),
-          { 
-            headers: { ...corsHeaders, 'Content-Type': 'application/json' },
-            status: 200
-          }
-        );
-      }
-    } catch (betaCheckError) {
-      console.error("Error checking beta user:", betaCheckError);
-      // Continue avec la vérification des abonnements même si la vérification beta échoue
+    // 2. Check for paid subscription (SECOND PRIORITY)
+    const paidResult = await checkPaidAccess(supabaseClient, user);
+    if (paidResult) {
+      return createResponse(paidResult);
     }
 
-    // Vérifier les abonnements payants - SECOND PRIORITAIRE
-    console.log("Checking paid subscription for user");
-    try {
-      const { data: paidSubscription, error: paidSubError } = await supabaseClient
-        .from('user_subscriptions')
-        .select('status, type, expires_at, promo_code')
-        .eq('user_id', user.id)
-        .eq('status', 'active')
-        .eq('type', 'paid')
-        .order('created_at', { ascending: false })
-        .limit(1)
-        .maybeSingle();
-      
-      if (paidSubError) {
-        console.error("Paid subscription check error:", paidSubError);
-      }
-      
-      if (paidSubscription) {
-        console.log('Abonnement payant trouvé:', paidSubscription);
-        
-        // Vérifier si l'abonnement est expiré
-        if (paidSubscription.expires_at) {
-          const expiryDate = new Date(paidSubscription.expires_at);
-          if (expiryDate < new Date()) {
-            console.log("Paid subscription expired at:", expiryDate);
-            return new Response(
-              JSON.stringify({ 
-                access: false, 
-                message: 'Abonnement expiré',
-                type: paidSubscription.type,
-                expires_at: paidSubscription.expires_at
-              }),
-              { 
-                headers: { ...corsHeaders, 'Content-Type': 'application/json' },
-                status: 200
-              }
-            );
-          }
-        }
-        
-        console.log("Active paid subscription found, granting access");
-        return new Response(
-          JSON.stringify({ 
-            access: true, 
-            type: paidSubscription.type,
-            expires_at: paidSubscription.expires_at,
-            promo_code: paidSubscription.promo_code
-          }),
-          { 
-            headers: { ...corsHeaders, 'Content-Type': 'application/json' },
-            status: 200
-          }
-        );
-      }
-    } catch (paidSubscriptionError) {
-      console.error("Error checking paid subscription:", paidSubscriptionError);
+    // 3. Check for trial subscription (LOWEST PRIORITY)
+    const trialResult = await checkTrialAccess(supabaseClient, user);
+    if (trialResult) {
+      return createResponse(trialResult);
     }
 
-    // Vérifier les abonnements d'essai - DERNIER PRIORITÉ
-    console.log("Checking trial subscription for user");
-    try {
-      const { data: trialSubscription, error: trialSubError } = await supabaseClient
-        .from('user_subscriptions')
-        .select('status, type, expires_at, promo_code')
-        .eq('user_id', user.id)
-        .eq('status', 'active')
-        .eq('type', 'trial')
-        .order('created_at', { ascending: false })
-        .limit(1)
-        .maybeSingle();
-      
-      if (trialSubError) {
-        console.error("Trial subscription check error:", trialSubError);
-      }
-      
-      if (trialSubscription) {
-        console.log('Abonnement d\'essai trouvé:', trialSubscription);
-        
-        // Vérifier si l'abonnement est expiré
-        if (trialSubscription.expires_at) {
-          const expiryDate = new Date(trialSubscription.expires_at);
-          if (expiryDate < new Date()) {
-            console.log("Trial subscription expired at:", expiryDate);
-            return new Response(
-              JSON.stringify({ 
-                access: false, 
-                message: 'Période d\'essai expirée',
-                type: trialSubscription.type,
-                expires_at: trialSubscription.expires_at
-              }),
-              { 
-                headers: { ...corsHeaders, 'Content-Type': 'application/json' },
-                status: 200
-              }
-            );
-          }
-        }
-        
-        console.log("Active trial subscription found, granting access");
-        return new Response(
-          JSON.stringify({ 
-            access: true, 
-            type: trialSubscription.type,
-            expires_at: trialSubscription.expires_at,
-            promo_code: trialSubscription.promo_code
-          }),
-          { 
-            headers: { ...corsHeaders, 'Content-Type': 'application/json' },
-            status: 200
-          }
-        );
-      }
-    } catch (trialSubscriptionError) {
-      console.error("Error checking trial subscription:", trialSubscriptionError);
-    }
-
-    // Aucun abonnement trouvé
+    // No subscription found
     console.log('Aucun abonnement actif trouvé pour', user.email);
-    return new Response(
-      JSON.stringify({ 
-        access: false, 
-        message: 'Aucun abonnement actif' 
-      }),
-      { 
-        headers: { ...corsHeaders, 'Content-Type': 'application/json' },
-        status: 200
-      }
-    );
+    return createResponse({ 
+      access: false, 
+      message: 'Aucun abonnement actif' 
+    });
 
   } catch (error) {
     console.error('Erreur générale:', error);
-    return new Response(
-      JSON.stringify({ 
-        error: error.message,
-        stack: error.stack,
-        access: false,
-        message: 'Erreur serveur lors de la vérification de l\'accès'
-      }),
-      { 
-        headers: { ...corsHeaders, 'Content-Type': 'application/json' },
-        status: 500
-      }
-    );
+    return createResponse({ 
+      error: error.message,
+      stack: error.stack,
+      access: false,
+      message: 'Erreur serveur lors de la vérification de l\'accès'
+    }, 500);
   }
 });
